--- src/main/java/com/example/flight/AirportNotFoundException.java
@@ -1,8 +1,0 @@
-package com.example.flight;
-
-public class AirportNotFoundException extends RuntimeException {
-
-    AirportNotFoundException(Long id) {
-        super("Could not found " + id);
-    }
-}
