--- src/main/java/com/example/flight/AirportRepository.java
@@ -1,6 +1,0 @@
-package com.example.flight;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-interface AirportRepository extends JpaRepository<Airport, Long> {
-}
