+++ src/main/java/com/example/flight/controllers/AirportController.java
@@ -1,7 +1,10 @@
-package com.example.flight;
+package com.example.flight.controllers;
 
 import java.util.List;
 
+import com.example.flight.entities.Airport;
+import com.example.flight.exceptions.AirportNotFoundException;
+import com.example.flight.repositories.AirportRepository;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
