+++ src/main/java/com/example/flight/entities/Airport.java
@@ -1,4 +1,4 @@
-package com.example.flight;
+package com.example.flight.entities;
 
 import lombok.*;
