+++ src/main/java/com/example/flight/exceptions/AirportNotFoundException.java
@@ -1,0 +1,8 @@
+package com.example.flight.exceptions;
+
+public class AirportNotFoundException extends RuntimeException {
+
+    public AirportNotFoundException(Long id) {
+        super("Could not found airport with ID: " + id);
+    }
+}
