+++ src/main/java/com/example/flight/repositories/AirportRepository.java
@@ -1,0 +1,7 @@
+package com.example.flight.repositories;
+
+import com.example.flight.entities.Airport;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface AirportRepository extends JpaRepository<Airport, Long> {
+}
